restructure test locaitons
