+++ src/spetlr/dbutils/__init__.py
@@ -1,0 +1,1 @@
+from .FileExists import file_exists  # noqa: F401
