+++ src/spetlr/testutils/__init__.py
@@ -1,0 +1,7 @@
+from .CleanupTestDatabases import CleanupTestDatabases  # noqa: F401
+from .stop_test_streams import stop_test_streams  # noqa: F401
+
+__all__ = [
+    stop_test_streams,
+    CleanupTestDatabases,
+]
