+++ tests/cluster/sql/test_sql_streaming.py
@@ -8,7 +8,7 @@ from spetlr.etl.loaders import SimpleLoader
 from spetlr.etl.loaders.stream_loader import StreamLoader
 from spetlr.spark import Spark
 from spetlr.sql import SqlHandle
-from spetlr.utils.stop_test_streams import stop_test_streams
+from spetlr.testutils.stop_test_streams import stop_test_streams
 from tests.cluster.sql.DeliverySqlServer import DeliverySqlServer
