+++ tests/local/utils/test_stop_test_streams.py
@@ -1,7 +1,7 @@
 from spetlrtools.testing import DataframeTestCase
 
 from spetlr.exceptions import OnlyUseInSpetlrDebugMode
-from spetlr.utils.stop_test_streams import stop_test_streams
+from spetlr.testutils.stop_test_streams import stop_test_streams
 
 
 class TestStopTestStreams(DataframeTestCase):
